(#156) refactor: integrate joinkm functionality into tablog with new join subcommand
- Removed the `joinkm` module and its associated tests, consolidating its functionality into the new `tablog join` subcommand.
- Updated `CHANGELOG.rst` to reflect the addition of the `tablog join` command for merging multiple KM markdown logs into a single time-sorted table.
- Modified `exec.py` to remove references to `joinkm` and include the new `tablog join` subparser.
- Enhanced `tablog.py` with the implementation of the `TropsTablogJoin` class, which handles input validation, file reading, sorting, and output writing.
- This refactor improves code organization and maintains consistency across command handling in the application.
