+++ src/trops/exec.py
@@ -14,7 +14,6 @@ from .trops import TropsCLI, TropsError
 from .utils import generate_sid
 from .view import add_view_subparsers
 from .tablog import add_tablog_subparsers
-from .joinkm import add_joinkm_subparsers
 
 
 def trops_git(args, other_args):
@@ -183,7 +182,6 @@ def main():
     add_ll_subparsers(subparsers)
     add_log_subparsers(subparsers)
     add_tablog_subparsers(subparsers)
-    add_joinkm_subparsers(subparsers)
     add_repo_subparsers(subparsers)
     add_view_subparsers(subparsers)
     add_show_subparsers(subparsers)
